Extract Nadia.Config module
